+++ colmet/common/backends/zeromq.py
@@ -3,6 +3,7 @@ ZeroMQ input backend for colmet collector
 '''
 
 import zmq
+from zmq.eventloop import ioloop, zmqstream
 import logging
 import socket
 import struct
@@ -19,36 +20,33 @@ class ZMQInputBackend(InputBaseBackend):
 
     def open(self):
         self.context = zmq.Context()
-        self.socket = self.context.socket(zmq.SUB)
-        self.socket.setsockopt(zmq.SUBSCRIBE, "")
+        self.loop = ioloop.IOLoop.instance()
+        self.socket = self.context.socket(zmq.PULL)
         self.socket.setsockopt(zmq.LINGER, self.options.zeromq_linger)
         self.socket.setsockopt(zmq.HWM, self.options.zeromq_hwm)
         self.socket.setsockopt(zmq.SWAP, self.options.zeromq_swap)
         LOG.debug("Use the bind URI '%s'" % self.options.zeromq_bind_uri)
         self.socket.bind(self.options.zeromq_bind_uri)
+        self.stream = zmqstream.ZMQStream(self.socket, self.loop)
 
     def close(self):
         self.socket.close()
         self.context.term()
 
-    def pull(self):
+    def unpack(self, msgs, callback):
         counters_list = []
-        try:
-            for i in xrange(1000):
-                raw = self.socket.recv(zmq.NOBLOCK, copy=False)
-                counters_list.extend(BaseCounters.unpack_to_list(raw.bytes))
-                del raw
-        except zmq.ZMQError, e:
-            if e.errno != zmq.EAGAIN:
-                raise e
+        for raw in msgs:
+            try:
+                counters_list.extend(BaseCounters.unpack_to_list(raw))
+            except Exception as e:
+                LOG.exception(e)
         LOG.debug("%s counters received" % len(counters_list))
-        if len(self.job_id_list) > 0:
-            counters_list = [metric for metric in counters_list
-                             if metric.job_id in self.job_id_list]
-            LOG.debug("%s counters received after filtering"
-                      % len(counters_list))
+        for counters in counters_list:
+            callback(counters)
 
-        return counters_list
+    def on_recv(self, callback):
+        intern_callback = lambda x: self.unpack(x, callback)
+        self.stream.on_recv(intern_callback)
 
 
 class ZMQOutputBackend(OutputBaseBackend):
@@ -60,7 +58,7 @@ class ZMQOutputBackend(OutputBaseBackend):
     def open(self):
         self.context = zmq.Context()
         self.hostname = socket.gethostname()
-        self.socket = self.context.socket(zmq.PUB)
+        self.socket = self.context.socket(zmq.PUSH)
         self.socket.setsockopt(zmq.LINGER, self.options.zeromq_linger)
         self.socket.setsockopt(zmq.HWM, self.options.zeromq_hwm)
         self.socket.setsockopt(zmq.SWAP, self.options.zeromq_swap)